Use String#isEmpty(), remove redundancies, and reduce verbose code (#1225)

+++ src/main/java/com/fasterxml/jackson/core/JsonParser.java
@@ -27,12 +27,12 @@ import com.fasterxml.jackson.core.util.RequestPayload;
 public abstract class JsonParser
     implements Closeable, Versioned
 {
-    private final static int MIN_BYTE_I = (int) Byte.MIN_VALUE;
+    private final static int MIN_BYTE_I = Byte.MIN_VALUE;
     // as per [JACKSON-804], allow range up to and including 255
-    private final static int MAX_BYTE_I = (int) 255;
+    private final static int MAX_BYTE_I = 255;
 
-    private final static int MIN_SHORT_I = (int) Short.MIN_VALUE;
-    private final static int MAX_SHORT_I = (int) Short.MAX_VALUE;
+    private final static int MIN_SHORT_I = Short.MIN_VALUE;
+    private final static int MAX_SHORT_I = Short.MAX_VALUE;
 
     /**
      * Enumeration of possible "native" (optimal) types that can be
@@ -1279,7 +1279,7 @@ public abstract class JsonParser
      * @since 2.8
      */
     public void finishToken() throws IOException {
-        ; // nothing
+        // nothing to do
     }
 
     /*
